+++ src/conversion/precompilePipeline.ts
@@ -4,14 +4,15 @@ import type { ParameterCommand } from '../commands/PARAMETER/ParameterCommand';
 import { sectionCommandParser } from '../commands/SECTION/sectionCommandParser';
 import { getParserForCommand } from '../commands/_common/getParserForCommand';
 import { parseCommand } from '../commands/_common/parseCommand';
-import type { $PipelineJson } from '../commands/_common/types/CommandParser';
-import type { $TaskJson } from '../commands/_common/types/CommandParser';
-import type { CommandBase } from '../commands/_common/types/CommandParser';
-import type { PipelineHeadCommandParser } from '../commands/_common/types/CommandParser';
-import type { PipelineTaskCommandParser } from '../commands/_common/types/CommandParser';
+import type {
+    $PipelineJson,
+    $TaskJson,
+    CommandBase,
+    PipelineHeadCommandParser,
+    PipelineTaskCommandParser,
+} from '../commands/_common/types/CommandParser';
 import { DEFAULT_TITLE } from '../config';
-import { ORDER_OF_PIPELINE_JSON } from '../constants';
-import { RESERVED_PARAMETER_NAMES } from '../constants';
+import { ORDER_OF_PIPELINE_JSON, RESERVED_PARAMETER_NAMES } from '../constants';
 import { ParseError } from '../errors/ParseError';
 import { UnexpectedError } from '../errors/UnexpectedError';
 import { HIGH_LEVEL_ABSTRACTIONS } from '../high-level-abstractions/index';
@@ -22,9 +23,7 @@ import type { TaskJson } from '../pipeline/PipelineJson/TaskJson';
 import type { PipelineString } from '../pipeline/PipelineString';
 import type { ScriptLanguage } from '../types/ScriptLanguage';
 import { SUPPORTED_SCRIPT_LANGUAGES } from '../types/ScriptLanguage';
-import type { number_integer } from '../types/typeAliases';
-import type { number_positive } from '../types/typeAliases';
-import type { string_name } from '../types/typeAliases';
+import type { number_integer, number_positive, string_name } from '../types/typeAliases';
 import { extractAllListItemsFromMarkdown } from '../utils/markdown/extractAllListItemsFromMarkdown';
 import { extractOneBlockFromMarkdown } from '../utils/markdown/extractOneBlockFromMarkdown';
 import { flattenMarkdown } from '../utils/markdown/flattenMarkdown';
@@ -61,6 +60,14 @@ export function precompilePipeline(pipelineString: PipelineString): PipelineJson
         knowledgePieces: [],
         personas: [],
         preparations: [],
+        sources: [
+            {
+                type: 'BOOK',
+                path: null,
+                // <- TODO: !!!!!! Pass here path of the file
+                content: pipelineString,
+            },
+        ],
     };
 
     function getPipelineIdentification() {
@@ -593,7 +600,6 @@ export function precompilePipeline(pipelineString: PipelineString): PipelineJson
 
     // =============================================================
 
-    // TODO: [🍙] Maybe do reorder of `$pipelineJson` here
     return exportJson({
         name: 'pipelineJson',
         message: `Result of \`precompilePipeline\``,